requirements.txt
cycler==0.10.0
decorator==4.4.0
Django==2.2.1
djangorestframework==3.9.3
kiwisolver==1.1.0
matplotlib==3.0.3
networkx==2.3
nose==1.3.7
numpy==1.16.3
pandas==0.24.2
py4j==0.10.7
pydot==1.4.1
pyparsing==2.4.0
pyspark-stubs==2.4.0.post1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
scipy==1.2.1
six==1.12.0
sqlparse==0.3.0
